closes #27
