+++ superpark-frontend/components/AppNavigator.js
@@ -1,4 +1,5 @@
-import React, { createContext, useContext, useState } from 'react';
+import React from 'react';
+import { Platform } from 'react-native';
 import { NavigationContainer } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
 
@@ -6,19 +7,39 @@ import { useAuth } from '../context/AuthContext';
 
 import DrawerScreens from './DrawerScreens';
 import LoginScreen from '../screens/Greeters/Login';
-import SignupPage from '../screens/Greeters/SignupPage_react';
+import SignupPage from '../screens/Greeters/SignupPage';
 import UserTypeSelection from '../screens/Greeters/UserTypeSelection';
 
-const Stack = createStackNavigator();
+const linking = {
+    prefixes: ['http://localhost:8081'],
+    config: {
+        screens: {
+            Drawer: 'dashboard',
+            'User Type': 'select-user',
+            Login: 'login',
+            'Sign Up': 'signup',
+        }
+    }
+}
+
+// function to check if login sesion is valid
+const validSession = () => {
+    if(Platform.OS === 'web') {
+        return (localStorage.getItem('token') != null)
+    } else {
+        return true;
+    }
+}
 
 const AppNavigator = () => {
     const { auth } = useAuth();
+    const Stack = createStackNavigator();
 
     return (
-        <NavigationContainer>
+        <NavigationContainer linking={linking}>
             <Stack.Navigator screenOptions={{headerShown: false}}>
                 {
-                    auth ? (
+                    (auth || validSession()) ? (
                         <Stack.Screen name="Drawer" component={DrawerScreens} />
                     ) : (
                         <Stack.Group>
